rename configurator.h to msp.h
